refactor: split index controller by feature
